format: format code

+++ src/quant/int8.py
@@ -7,13 +7,15 @@ _EPS = 1e-12
 _warned_shape_fallback = False
 
 
-def quantize_int8(x: torch.Tensor, dim: int | None = None) -> tuple[torch.Tensor, torch.Tensor]:
+def quantize_int8(
+    x: torch.Tensor, dim: int | None = None
+) -> tuple[torch.Tensor, torch.Tensor]:
     if dim is None:
         abs_max = x.detach().abs().max().float()
     else:
         abs_max = x.detach().abs().amax(dim=dim, keepdim=True).float()
     scale = (abs_max / _INT8_MAX).clamp_min(_EPS)
-    x_i8 =  torch.round(x.float() / scale).clamp(-_INT8_MAX, _INT8_MAX).to(torch.int8)
+    x_i8 = torch.round(x.float() / scale).clamp(-_INT8_MAX, _INT8_MAX).to(torch.int8)
     return x_i8, scale
 
 
@@ -38,9 +40,9 @@ def int8_gemm(a, b, out_dtype, rowwise=False):
                 "the fake-quant fallback (no int8 kernel speedup). Warned once."
             )
             _warned_shape_fallback = True
-        return (
-            fake_quantize_int8(a, dim=a_dim) @ fake_quantize_int8(b, dim=b_dim)
-        ).to(out_dtype)
+        return (fake_quantize_int8(a, dim=a_dim) @ fake_quantize_int8(b, dim=b_dim)).to(
+            out_dtype
+        )
 
     a_i8, a_scale = quantize_int8(a, dim=a_dim)  # rowwise: (M,1); else scalar
     b_i8, b_scale = quantize_int8(b, dim=b_dim)  # rowwise: (1,N); else scalar
